pyproject.toml
[tool.poetry]
name = "store"
version = "0.1.0"
description = ""
authors = ["Евгений Нестеренко <hectep@gdematerial.ru>"]

[tool.poetry.dependencies]
python = "^3.6"
django = "2.2.7"
djangorestframework = "^3.11.0"
celery = "^4.4.2"
requests = "^2.23.0"
pytest-django = "^3.9.0"
mixer = "^6.1.3"
psycopg2 = "^2.8.5"
redis = "^3.5.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
